[chore] product_list 폴더 js파일 모듈화
