+++ client/css/product-detail.css
@@ -518,7 +518,7 @@
 .product-review_list-notice_text {
   display: none;
   background: var(--gray-50);
-  ont-weight: 600;
+  font-weight: 600;
   font-size: 12.003px;
   color: var(--content);
   padding: 20px;
