+++ server/db/data.json
@@ -105,7 +105,7 @@
   "basket": [
     {
       "id": "product-rksk",
-      "number": 25
+      "number": 31
     },
     {
       "id": "product-ekfk",
